Use one shared author-identity check across every tab
The Co-author Network tab still showed a user her own maiden name as
her most frequent collaborator, because author identity was implemented
four separate times and only the narrative path had been fixed. The
citation network, the narrative's author linkification and the
co-author geo service each compared author strings with ===, and the
citation network additionally guessed the profile owner from byline
frequency instead of being told who it is.

Extracts the identity logic into src/utils/authorIdentity.ts as the
single definition - isSameResearcher, coAuthorsOf, inferMainAuthor -
and routes all four call sites through it. CitationNetwork now receives
the owner's name as a prop, and solo-paper detection matches on
identity so a paper authored solely under a name variant counts as solo.

On the reported profile: top collaborator MK Pein (25 papers) becomes
J Breitkreutz (12), unique co-authors 113 to 111, and no owner variant
appears as a collaborator in any tab. Matrix 55 passing, 0 regressions;
all 326 cached profiles sweep clean.
